fix(inventory): make read-side Dapper queries run against PostgreSQL
Three latent defects surfaced by the new tenant-isolation integration
test (all in the read-side StockRead slice), none previously exercised
against a live engine:

- Dapper 2.1.x does not bind DateOnly/TimeOnly natively, so every query
  passing a DateOnly parameter (expiry @Today, consumption @From/@To)
  threw NotSupportedException. Register DateOnly/TimeOnly type handlers
  once at composition (AddSislabInfrastructure).
- PostgreSQL COUNT returns bigint; the read models expose the counts as
  int, so Dapper could not match the record constructors. Cast every
  projected COUNT (KPIs, per-location counts, COUNT(*) OVER() totals,
  movement counts) to ::int.
- Alias the paginated total_rows column as totalrows so Dapper binds it
  to the TotalRows record parameter (underscore names are not matched).
- The consumption report ran two statements in one round-trip sharing a
  WITH consumption CTE, but a CTE is scoped to a single statement, so
  the second SELECT failed with relation consumption does not exist.
  Prepend the CTE to both statements.

+++ src/Modules/Inventory/SISLAB.Modules.Inventory.Application/StockRead/GetBelowMinimumSummaryQuery.cs
@@ -40,7 +40,7 @@ internal sealed class GetBelowMinimumSummaryQueryHandler
     // mandatory tenant scoping (the read side has no EF global query filter).
     private const string Sql =
         """
-        SELECT COUNT(*) AS belowminimumcount
+        SELECT COUNT(*)::int AS belowminimumcount
         FROM inventory.stock_view AS v
         WHERE v.company_id = @CompanyId
           AND v.is_below_minimum;
